app/config.py
import os

from dotenv import load_dotenv

load_dotenv()

PG_USER = os.getenv("POSTGRES_USER")
PG_PASS = os.getenv("POSTGRES_PASSWORD")
PG_DB = os.getenv("POSTGRES_DB")
HOST = os.getenv("POSTGRES_HOST")
PORT = os.getenv("POSTGRES_PORT")
SECRET = os.getenv("SECRET_KEY")
