app/requirements.txt
aioredis==1.3.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
idna==3.10
makefun==1.15.6
Mako==1.3.9
MarkupSafe==3.0.2
pendulum==3.0.0
# psycopg2==2.9.10
psycopg2-binary==2.9.10
pwdlib==0.2.1
pybase62==1.0.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==4.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
time-machine==2.16.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.0
